general: tried to clean up the validation rule argument workaround
